refactor(proto): remove transform module and SDK dependency
