+++ opentelemetry-otlp/src/exporter/tonic/logs.rs
@@ -152,6 +152,7 @@ impl LogExporter for TonicLogsClient {
     }
 
     fn set_resource(&mut self, resource: &opentelemetry_sdk::Resource) {
-        self.resource = crate::transform::common::tonic::resource_to_attributes_with_schema(resource);
+        self.resource =
+            crate::transform::common::tonic::resource_to_attributes_with_schema(resource);
     }
 }
