+++ opentelemetry-otlp/src/exporter/tonic/trace.rs
@@ -1,11 +1,11 @@
 use core::fmt;
 use std::sync::{Arc, Mutex};
 
+use crate::transform::trace::tonic::group_spans_by_resource_and_scope;
 use opentelemetry::{otel_debug, otel_warn};
 use opentelemetry_proto::tonic::collector::trace::v1::{
     trace_service_client::TraceServiceClient, ExportTraceServiceRequest,
 };
-use crate::transform::trace::tonic::group_spans_by_resource_and_scope;
 use opentelemetry_sdk::error::OTelSdkError;
 use opentelemetry_sdk::{
     error::OTelSdkResult,
@@ -157,6 +157,7 @@ impl SpanExporter for TonicTracesClient {
     }
 
     fn set_resource(&mut self, resource: &opentelemetry_sdk::Resource) {
-        self.resource = crate::transform::common::tonic::resource_to_attributes_with_schema(resource);
+        self.resource =
+            crate::transform::common::tonic::resource_to_attributes_with_schema(resource);
     }
 }
