--- opentelemetry-proto/src/transform/mod.rs
@@ -1,16 +1,0 @@
-pub mod common;
-
-#[cfg(feature = "metrics")]
-pub mod metrics;
-
-#[cfg(feature = "trace")]
-pub mod trace;
-
-#[cfg(feature = "logs")]
-pub mod logs;
-
-#[cfg(feature = "zpages")]
-pub mod tracez;
-
-#[cfg(feature = "profiles")]
-pub mod profiles;
