--- opentelemetry-proto/src/transform/profiles.rs
@@ -1,1 +1,0 @@
-
